SBERDOMA-1012 remove access schema
