+++ apps/condo/domains/meter/access/Meter.js
@@ -2,32 +2,92 @@
  * Generated by `createschema meter.Meter 'number:Text; billingAccountMeter?:Relationship:BillingAccountMeter:SET_NULL; account?:Relationship:BillingAccount:SET_NULL; property:Relationship:Property:CASCADE; unitName:Text; place?:Text; resource:Relationship:MeterResource:CASCADE;'`
  */
 
-const { RESIDENT } = require('@condo/domains/user/constants/common')
-const { canReadMeterEntity, canManageMeterEntity } = require('../utils/accessSchema')
 const { Meter } = require('../utils/serverSchema')
 const { throwAuthenticationError } = require('@condo/domains/common/utils/apolloErrorFormatter')
+const { Property } = require('@condo/domains/property/utils/serverSchema')
+const get = require('lodash/get')
+const { checkPermissionInUserOrganizationOrRelatedOrganization } = require('../../organization/utils/accessSchema')
+const { RESIDENT } = require('@condo/domains/user/constants/common')
+const { queryOrganizationEmployeeFromRelatedOrganizationFor } = require('@condo/domains/organization/utils/accessSchema')
+const { queryOrganizationEmployeeFor } = require('@condo/domains/organization/utils/accessSchema')
 
 async function canReadMeters ({ authentication: { item: user } }) {
     if (!user) return throwAuthenticationError()
-    return await canReadMeterEntity({ user })
+    if (user.isAdmin) return {}
+
+    if (user.type === RESIDENT) {
+        return {
+            createdBy: { id: user.id },
+        }
+    }
+    const userId = user.id
+    return {
+        organization: {
+            OR: [
+                queryOrganizationEmployeeFor(userId),
+                queryOrganizationEmployeeFromRelatedOrganizationFor(userId),
+            ],
+        },
+    }
 }
 
 async function canManageMeters ({ authentication: { item: user }, originalInput, operation, itemId, context }) {
     if (!user) return throwAuthenticationError()
+    if (user.isAdmin) return true
 
     if (operation === 'update' && user.type === RESIDENT) {
         const [meter] = await Meter.getAll(context, { id: itemId, createdBy: { id: user.id } })
         return !!meter
     }
 
-    return await canManageMeterEntity({
-        schema: Meter,
-        user,
-        itemId,
-        operation,
-        originalInput,
-        context,
-    })
+    if (operation === 'create') {
+        const organizationIdFromMeter = get(originalInput, ['organization', 'connect', 'id'])
+        if (!organizationIdFromMeter) {
+            return false
+        }
+
+        const propertyId = get(originalInput, ['property', 'connect', 'id'])
+        const [property] = await Property.getAll(context, { id: propertyId })
+        if (!property) {
+            return false
+        }
+
+        if (user.type === RESIDENT) {
+            return true
+        }
+
+        const organizationIdFromProperty = get(property, ['organization', 'id'])
+        if (organizationIdFromMeter !== organizationIdFromProperty)
+            return false
+
+        return await checkPermissionInUserOrganizationOrRelatedOrganization(context, user.id, organizationIdFromMeter, 'canManageMeters')
+
+    } else if (operation === 'update') {
+        if (!itemId) {
+            return false
+        }
+
+        const [meter] = await Meter.getAll(context, { id: itemId })
+        if (!meter)
+            return false
+
+        // if we pass property then we need check that this Property is in the organization in which the Meter is located
+        const organizationIdFromMeter = get(meter, ['organization', 'id'])
+        const propertyId = get(originalInput, ['property', 'connect', 'id'])
+        if (propertyId) {
+            const [property] = await Property.getAll(context, { id: propertyId })
+            if (!property)
+                return false
+
+            const organizationIdFromProperty = get(property, ['organization', 'id'])
+            if (organizationIdFromMeter !== organizationIdFromProperty)
+                return false
+        }
+
+        return await checkPermissionInUserOrganizationOrRelatedOrganization(context, user.id, organizationIdFromMeter, 'canManageMeters')
+    }
+
+    return false
 }
 
 /*
